+++ Client/src/components/ContactUs/ContactUs.css
@@ -1,0 +1,3 @@
+.contact-us{
+    height:100%;
+}
