+++ Client/src/components/ContactUs/ContactUs.jsx
@@ -1,23 +1,14 @@
-import React, { useState } from "react";
-
+import React from "react";
+import './ContactUs.css'
 
 export default function ContactUs() {
-    const [showDescription, setShowDescription] = useState(false)
 
-    const toggleDescription = () => {
-        setShowDescription(!showDescription)
-    };
     return (
         <div className="Contactanos">
-            <button onClick={toggleDescription}>Contáctanos</button>
-            {
-                showDescription && (
-                    <div className="contac-us">
-                    <h2>Contáctanos</h2>
-                    <p>Puedes contactarnos en Whatsapp al 1121423423</p>
-                    </div>
-                )
-            }
+            <div className="contac-us">
+                <h2>Contáctanos</h2>
+                <p>Puedes contactarnos en Whatsapp al 1121423423</p>
+            </div>
         </div>
 
     )
